Add util module cfgext
